renamed Transafer folder to Transfer, added Transfer.scss and moved all styles to it, implemented BEM
